refactor: Move hooks folder too root
